코드 리뷰 반영 리팩토링 및 기타 수정
