+++ client/src/component/mainpage/Banner.js
@@ -44,12 +44,10 @@ function Banner() {
   return (
     <StyledCarousel autoplay={true} animation={'slide'} timeout={100}>
       {items.map((item, idx) => (
-        <>
-          <MainBanner key={'banner' + idx}>
-            <BannerImg src={item.src}></BannerImg>
-            <CurrentPage>{idx + 1 + '/' + items.length}</CurrentPage>
-          </MainBanner>
-        </>
+        <MainBanner key={`banner-${idx}`}>
+          <BannerImg src={item.src}></BannerImg>
+          <CurrentPage>{idx + 1 + '/' + items.length}</CurrentPage>
+        </MainBanner>
       ))}
     </StyledCarousel>
   );
