+++ client/src/component/share/Product.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-
-import ProductItem from './ProductItem';
+import ProductItem from 'component/share/ProductItem';
 
 const ProductContainerHeader = styled.div`
   font-family: 'BMDOHYEON';
