+++ client/src/component/share/util.js
@@ -1,0 +1,3 @@
+export function addCommaToNumber(x) {
+  return x.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ',');
+}
